src/requirements.txt
azure-identity==1.15.0
azure-mgmt-alertsmanagement==1.0.0
fastapi==0.110.0
fastapi-health==0.4.0
fastapi-restful==0.5.0
Jinja2==3.1.3
jinja2-time==0.2.0
jinja2-humanize-extension==0.4.0
json-logging==1.3.0
openpyxl==3.1.2
pandas==2.2.1
prometheus-api-client==0.5.5
pydantic==2.6.3
pydantic-settings==2.2.1
pytz==2024.1
PyYAML==6.0.1
starlette>=0.36.3,<0.37.0  # pin to <0.37 (fastapi==0.110.0)
starlette_exporter==0.21.0
typing-inspect==0.9.0
uvicorn==0.28.0
